backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI Services
google-generativeai>=0.8.0
google-genai>=0.4.0

# Image handling
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Web scraping (for Realtor.ca listing scraper)
playwright>=1.40.0

